+++ Modules/Intent.Modules.Application.FluentValidation/NuGetPackages.cs
@@ -1,0 +1,9 @@
+﻿using Intent.Modules.Common.VisualStudio;
+
+namespace Intent.Modules.Application.FluentValidation
+{
+    public class NuGetPackages
+    {
+        public static readonly INugetPackageInfo FluentValidationDependencyInjectionExtensions = new NugetPackageInfo("FluentValidation.DependencyInjectionExtensions", "11.6.0");
+    }
+}