+++ Modules/Intent.Modules.Blazor.HttpClients/Templates/DtoContract/DtoContractTemplatePartial.cs
@@ -1,26 +1,12 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Intent.Application.FluentValidation.Api;
 using Intent.Engine;
-using Intent.Metadata.Models;
-using Intent.Modelers.Domain.Api;
 using Intent.Modelers.Services.Api;
-using Intent.Modelers.Types.ServiceProxies.Api;
 using Intent.Modules.Blazor.HttpClients.Templates.EnumContract;
 using Intent.Modules.Blazor.HttpClients.Templates.PagedResult;
-using Intent.Modules.Common;
-using Intent.Modules.Common.CSharp.Builder;
-using Intent.Modules.Common.CSharp.Templates;
-using Intent.Modules.Common.Templates;
-using Intent.Modules.Common.TypeResolution;
 using Intent.Modules.Contracts.Clients.Shared.Templates.DtoContract;
 using Intent.RoslynWeaver.Attributes;
-using Intent.SdkEvolutionHelpers;
-using Intent.Templates;
-using Intent.Utils;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: DefaultIntentManaged(Mode.Ignore, Targets = Targets.Usings)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
 
 namespace Intent.Modules.Blazor.HttpClients.Templates.DtoContract