+++ Tests/Standard.AspNetCore.TestApplication/Standard.AspNetCore.TestApplication.Api/Controllers/ValidationTestingController.cs
@@ -1,0 +1,46 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Intent.RoslynWeaver.Attributes;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Standard.AspNetCore.TestApplication.Application;
+using Standard.AspNetCore.TestApplication.Application.Interfaces;
+using Standard.AspNetCore.TestApplication.Application.Validation;
+using Standard.AspNetCore.TestApplication.Domain.Common.Interfaces;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: DefaultIntentManaged(Mode.Fully, Targets = Targets.Usings)]
+[assembly: IntentTemplate("Intent.AspNetCore.Controllers.Controller", Version = "1.0")]
+
+namespace Standard.AspNetCore.TestApplication.Api.Controllers
+{
+    [ApiController]
+    [Route("api/validation")]
+    public class ValidationTestingController : ControllerBase
+    {
+        private readonly IValidationTestingService _appService;
+        private readonly IUnitOfWork _unitOfWork;
+
+        public ValidationTestingController(IValidationTestingService appService, IUnitOfWork unitOfWork)
+        {
+            _appService = appService ?? throw new ArgumentNullException(nameof(appService));
+            _unitOfWork = unitOfWork ?? throw new ArgumentNullException(nameof(unitOfWork));
+        }
+
+        /// <summary>
+        /// </summary>
+        /// <response code="200">Returns the specified ValidationDto.</response>
+        [HttpGet("validation-operation")]
+        [ProducesResponseType(typeof(ValidationDto), StatusCodes.Status200OK)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult<ValidationDto>> ValidationOperation(CancellationToken cancellationToken = default)
+        {
+            var result = default(ValidationDto);
+            result = await _appService.ValidationOperation(cancellationToken);
+            return Ok(result);
+        }
+    }
+}