+++ Modules/Intent.Modules.Blazor.HttpClients.Dtos.FluentValidation/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 1.0.1
 
+- Improvement: Added support for the [new Email Address](https://github.com/IntentArchitect/Intent.Modules.NET/blob/development/Modules/Intent.Modules.Application.FluentValidation/release-notes.md#version-383) option on the Validation stereotype.
 - Fixed: A Software Factory crash around `DTO Unique Constraint validation` for DTO's mapping to constructors / operations.
 
 ### Version 1.0.0